+++ hypergan/__init__.py
@@ -1,6 +1,7 @@
 import hypergan
 from .gan import GAN
 from .graph import Graph
+from .config import selector
 import tensorflow as tf
 import hypergan.cli
