+++ hypergan/graph/__init__.py
@@ -1,0 +1,2 @@
+from .common import *
+from .graph import *
