+++ hypergan/graph/common.py
@@ -1,0 +1,10 @@
+
+def sigmoid_kl_with_logits(logits, targets):
+   # broadcasts the same target value across the whole batch
+   # this is implemented so awkwardly because tensorflow lacks an x log x op
+   assert isinstance(targets, float)
+   if targets in [0., 1.]:
+     entropy = 0.
+   else:
+     entropy = - targets * np.log(targets) - (1. - targets) * np.log(1. - targets)
+     return tf.nn.sigmoid_cross_entropy_with_logits(logits, tf.ones_like(logits) * targets) - entropy
