+++ hypergan/graph/graph.py
@@ -323,3 +323,13 @@ class Graph:
         g_cost, d_fake_cost, d_real_cost = sess.run([g_loss, d_fake, d_real])
 
         return g_cost,d_fake_cost, d_real_cost,0
+
+def sigmoid_kl_with_logits(logits, targets):
+   # broadcasts the same target value across the whole batch
+   # this is implemented so awkwardly because tensorflow lacks an x log x op
+   assert isinstance(targets, float)
+   if targets in [0., 1.]:
+     entropy = 0.
+   else:
+     entropy = - targets * np.log(targets) - (1. - targets) * np.log(1. - targets)
+     return tf.nn.sigmoid_cross_entropy_with_logits(logits, tf.ones_like(logits) * targets) - entropy
