+++ hypergan/trainers/__init__.py
@@ -1,0 +1,5 @@
+from .common import *
+from . import adam_trainer
+from . import slowdown_trainer
+from . import sgd_adam_trainer
+from . import rmsprop_trainer
