+++ hypergan/trainers/adam_trainer.py
@@ -1,6 +1,7 @@
 import tensorflow as tf
 import numpy as np
 from hypergan.util.globals import *
+from .common import *
 
 def initialize(config, d_vars, g_vars):
     d_loss = get_tensor('d_loss')
