+++ hypergan/util/__init__.py
@@ -1,0 +1,3 @@
+from .globals import *
+from .ops import *
+from .hc_tf import *
